requirements.txt
python-telegram-bot==20.7
flask==3.0.0
requests==2.31.0
steampy==1.3.0
steam==1.4.4
